requirements.txt
asgiref==3.4.1
Django==3.2.8
powerline-status==2.7
PyMySQL==1.0.2
pytz==2021.3
sqlparse==0.4.2

opencv-python~=4.5.5.64
numpy~=1.22.3
pandas~=1.4.2
Pillow~=9.1.0
cnocr~=2.1.0
self~=2020.12.3